Move kernel connector to notebooks
